backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
feedparser==6.0.10
nltk==3.8.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0
